requirements.txt
click==7.1.2
Flask==1.1.2
identify==2.2.2
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
python-magic-bin==0.4.14
Werkzeug==1.0.1
